AI_VisionAssistant/requirements.txt
google-ai-generativelanguage==0.6.10
google-generativeai==0.8.3
gTTS==2.5.4
langchain-core==0.3.21
langchain-google-genai==2.0.6
langsmith==0.1.147
pillow==11.0.0
streamlit==1.40.2